hentAlle i meldekortservice returnerer gjeldende vedtak og alle meldekort
